+++ src/components/initiativeTable/CreatureTable.svelte
@@ -8,7 +8,7 @@
     import CreatureRowUnlocked from "../creature/unlocked/CreatureRowUnlocked.svelte";
     import Speed from "../../graphics/icons/speed.svelte";
     import type { EntityType } from "../../lib/models/entity";
-    import { CREATURE_TYPES } from "../../lib/typeFilters";
+    import { CREATURE_TYPES } from "../../lib/models/typeFilters";
     
     export let typeFilter: EntityType[];
     $: creatures = $entities.filter(e => CREATURE_TYPES.includes(e.type) && typeFilter.includes(e.type));
