--- src/lib/keys.ts
@@ -1,8 +1,0 @@
-export const KEYS = [
-    "Tab",
-    "Enter",
-    "ArrowUp",
-    "ArrowDown",
-    "ArrowLeft",
-    "ArrowRight"
-]
