+++ src/lib/models/focusType.ts
